+++ typescript/infra/scripts/check/check-utils.ts
@@ -23,10 +23,7 @@ import { eqAddress, objFilter } from '@hyperlane-xyz/utils';
 
 import { Contexts } from '../../config/contexts.js';
 import { DEPLOYER } from '../../config/environments/mainnet3/owners.js';
-import {
-  getWarpAddresses,
-  getWarpAddressesFromMergedRegistry,
-} from '../../config/registry.js';
+import { getWarpAddressesFrom } from '../../config/registry.js';
 import { getWarpConfig } from '../../config/warp.js';
 import { chainsToSkip } from '../../src/config/chain.js';
 import { DeployEnvironment } from '../../src/config/environment.js';
@@ -207,9 +204,8 @@ export async function getGovernor(
       warpRouteId,
       registryUris,
     );
-    const warpAddresses = registryUris
-      ? await getWarpAddressesFromMergedRegistry(warpRouteId, registryUris)
-      : getWarpAddresses(warpRouteId);
+    const warpAddresses = await getWarpAddressesFrom(warpRouteId, registryUris);
+
     const filteredAddresses = Object.keys(warpAddresses) // filter out changes not in config
       .filter((key) => key in config)
       .reduce((obj, key) => {