+++ typescript/ccip-server/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/ccip-server
 
+## 12.2.0
+
 ## 12.1.0
 
 ## 12.0.0