+++ typescript/github-proxy/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/github-proxy
 
+## 12.2.0
+
 ## 12.1.0
 
 ## 12.0.0